Move all services to service folder (#147)
