+++ azure-pipelines.yml
@@ -13,7 +13,7 @@ jobs:
   steps:
   - task: NodeTool@0
     inputs:
-      versionSpec: '10.9'
+      versionSpec: '10.13'
     displayName: 'Install Node.js'
 
   - script: |
