+++ src/app/button-page/button-page.component.spec.ts
@@ -15,8 +15,8 @@ along with OVFPlayer.  If not, see <https://www.gnu.org/licenses/>.
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ButtonPageComponent } from './button-page.component';
-import { BoardService } from '../board.service';
-import { SpeechbarService } from '../speechbar.service';
+import { BoardService } from '../services/board/board.service';
+import { SpeechbarService } from '../services/speechbar/speechbar.service';
 import { of } from 'rxjs';
 import { OBFBoard, Button } from '../obfboard';
 import { ObfButtonComponent } from '../obf-button/obf-button.component';
@@ -24,7 +24,7 @@ import { MatRippleModule } from '@angular/material';
 import { ProgressComponent } from '../progress/progress.component';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { InteractionEventHandlerDirective } from '../interaction-event-handler.directive';
-import { ConfigService, InteractionEventType } from '../config.service';
+import { ConfigService, InteractionEventType } from '../services/config/config.service';
 
 describe('ButtonPageComponent', () => {
   let component: ButtonPageComponent;
